refactor loading/merging files
