+++ tests/loading/test_field_merges.py
